Added spaces for readability.
Changed the crazy xpath find methods to be simple
click_links. This works because the click_link
method will also search the alt text of an
image.

I didn’t add factories to this at the moment because
it’s doing a larger process of signing up a new user